+++ src/index.scss
@@ -7,8 +7,8 @@
 @import 'Components/Common/NavBar';
 @import 'Components/Common/Jumbotron';
 @import 'Components/Common/Loader';
+@import 'Components/Form/NoStyleButton';
 @import 'Components/Form/TextField';
-@import 'Components/Form/Button/NoStyleButton';
 @import 'Components/Thumbnail';
 @import 'Components/Sign/SignInForm';
 @import 'Components/Sign/SignUpForm';
